Another sample file
This is the same program as in.txt except formatted to occupy multiple lines. Also, it is formatted in such a way that there is no whitespace at the end of a line. The line break is added in the lexer program since it cannot be added into the actual text file.