Node Class and Tree Serialization
This features the node class we will be using for this project and the functions to use for later. We can add/change some of its attributes to better coincide with our lexer output. Also, functions for traversing through the tree are included here.

The main() function is a showcase of the functions that precede it.